feat: bump BINFILE_SIGNATURE and add portable I/O primitives
Version gate for the breaking .bin format change that follows.

Replaced the DUMP_BIN/RELOAD_BIN macros with write_raw/read_raw/
write_raw_as/read_raw_as/dump_bin/reload_bin templates (util/binfile.h).

Add a dedicated write_raw_long32/read_raw_long32 for the one narrowing
conversion here with real overflow risk (long, 8 bytes on Linux, to
int32_t).

Converted snapglob.cpp's 20 call sites.

+++ src/snaplib/snapdata/survdata.h
@@ -133,7 +133,7 @@ typedef struct           /* Data relating to an observation or set of obs */
         vecdata *vdata;
         pntdata *pdata;
     } obs;
-    long obsalloc;      /* Space allocated for the observations */
+    int32_t obsalloc;   /* Space allocated for the observations */
     int maxcvr;       /* Allocated size of covariance matrices */
     int maxclass;     /* Allocated size of classification array */
     int maxsyserr;    /* Allocated size of systematic error array */
